refactor(metrics): move reward subscores/geometry/config into diffusion_planner.metrics
PR1 for #130. Pure, byte-identical move of the raw subscore computation,
geometry primitives, and RewardConfig out of rlvr/reward.py into a neutral
diffusion_planner.metrics package, so the base-SFT validation loop can import
them without depending on rlvr. No logic change.

rlvr/reward.py keeps the reward shaping (RewardBreakdown, compute_reward_batch,
compute_reward, compute_group_advantages) and re-exports every moved symbol via
`from diffusion_planner.metrics.* import *`, so `from rlvr.reward import ...`
keeps working for all 27 externally-imported symbols.

Verified: all 31 moved def/class sources byte-identical (ast), golden tests and
the full reward-importing suite green, no behavior change (guaranteed by PR0
goldens).
