+++ diffusion_planner/diffusion_planner/metrics/__init__.py
@@ -1,0 +1,13 @@
+"""Neutral shared metrics library (issue #130).
+
+Raw EPDMS-style subscores + geometry + config, carved out of ``rlvr.reward`` so
+the base-SFT validation loop can import them without depending on ``rlvr``.
+``rlvr.reward`` re-exports the same symbols for backward compatibility.
+"""
+
+from __future__ import annotations
+
+from diffusion_planner.metrics.config import *  # noqa: F401,F403
+from diffusion_planner.metrics.config import RewardConfig  # noqa: F401
+from diffusion_planner.metrics.geometry import *  # noqa: F401,F403
+from diffusion_planner.metrics.subscores import *  # noqa: F401,F403
